scripts/cargo: Never do anything with librusty_v8.a until it is finished downloading
This is to protect against build race conditions we were seeing in debug builds on highly parallel
machines.  Specifically, multiple instances of the script were entering the binding.o conditional
and attempting to concurrently manipulate the library's symbols.  It's unclear whether this ever
actually corrupted the file, but it was sometimes causing one of the subprocesses to die with a bus
error.